add final refactoring for review
