+++ src/main/java/by/photoshop/smelnik/api/formsParameters/Parameters.java
@@ -1,4 +1,4 @@
-package by.photoshop.smelnik.api_tests.utils.forms;
+package by.photoshop.smelnik.api.formsParameters;
 
 import com.github.javafaker.Faker;
